requirements.txt
pandas==1.2.4
numpy==1.20.2
scikit-learn==0.24.2
nltk==3.6.2
gradio==2.0.8
pyyaml==5.4.1